Convert tests to ansible-test
pytest was just a placeholder while we waited for ansible-test to grow
collection support, which it has now done. This support is still in
flux, but the bits we want (integration, pep8) appear to be working.

While in the top level of this working tree, run
`<path_to_ansible_checkout>/bin/ansible-test integration`

This completely fails under Python 3 (very weird, haven't had time
to track down why it can't find the modules) and can encounter
interesting unicode problems under Python 2, but It Works On My
Machine(TM).