Make test code python 2 and 3 compatible
There were two main issues with the code that needed fixing in order
to make it python 2 compatible:

 1. unittest.mock in not part of the standard library on python 2 and
 2. python 2 uses relative imports if not instructed otherwise.

We solved the first issue by placing the mock imports into the utils
module and making sure that we handle python 2 as well.

Second issue was solved by adding absolute_import to the top of the
source files, making imports behave like in python 3. And while we
were at it, we also added other future imports that Ansible linter
enforces and set the metaclass accordingly.