Shorten stdout in when unit test fails
When a specific test fails, pytest would print out entire failing
test case, together with entire `.parametrize()` stanza verbatim as
provided in the test file. But this is quite unreadable since our
matrix (that we feed to the `.parametrize()` stanza) is quite big.

With this commit we therefore extract the matrix definition into its
own variable so that when a specific test fails only short section is
printed to stdout, without matrix content. Namely:

```
@pytest.mark.parametrize('test_data', matrix, ids=generate_name)
    def test_module(self, test_data):
        self.run_test_case(test_data)
```

Signed-off-by: Miha Pleško <miha.plesko@xlab.si>