Make asset modules use new utilities and HTTP client
The benefit of this move is simplified testing of module
functionality. It is now almost trivial to validate module parameters
to payload transformation.

We can also get rid of the tests that are not concerned with parameter
parsing since synchronization and HTTP access are already tested by a
separate battery of tests.