The sensu.sensu_go.install role
With this commit we introduce a role that configures
target's package manager and then installs desired Sensu Go
binary with it. Following binaries are currently present:

- sensu-go-backend
- sensu-go-agent
- sensu-go-cli

Currently the role supports following target platforms:

- Centos 6 & 7
- Ubuntu 14.04 & 16.04 & 18.04 & 18.10 & 19.04

Role allows you to either install latest version of the
component or specify specific version, but that version
must exist in the stable package repository.

Signed-off-by: Miha Pleško <miha.plesko@xlab.si>