Move test suite to installed ansible-test command
Now that beta release of Ansible 2.9 is out, we can start running our
tests against a released version.

But because ansible-test command is not yet completely independent from
the core ansible plugins, we removed the --requirements switch from the
test commands. Instead, we created two pip requirement files that
contain just the packages we need for our tests to pass.

This has an additional side-effect of considerably reducing the number
of packages that need to be installed when running tests.