Introduce simple debug logging
With this commit we introduce a bit hackish yet
useful debug logging for modules. Logging is disabled
by default, you need to specify env variable in the
playbook to turn it on:

```
environment:
  SENSU_ANSIBLE_DEBUG: yes
```

Then the log file /tmp/sensu-ansible.log will appear
on target host. For now we only log all the API requests
(with payloads) because that's what we usuall need to
see.

Signed-off-by: Miha Pleško <miha.plesko@xlab.si>