Rework CircleCI configuration into a test matrix
CircleCI does not natively support matrix testing. This is why we had
to write our own version of it. The way the CircleCI config is
structured now is:

 1. Workflows section describe the test matrix on a high level (this
    is primarily meant for human consumption). This is also the reason
    why we moved this section to the top of the file.
 2. Jobs section is the first part of the "matrix" implementation.
    Version of Ansible is specified here via environment variable.
 3. Executors section is the second part of the matrix implementation
    and controls the python version that is used for running tests.

Commands section contains the reusable operation definitions that are
used by the jobs section. Note that we removed all of the parameters
from the commands and replaced them with environment variables. We did
this to unify the argument passing. Previously, we were using the mix
of variables and parameters. We went for the variables because they
are also supported by the executors section.