Initial implementation of the sensu.sensu_go.backend role
The role is supposed to be run against nodes where we want to
have the Sensu backend running. It installs the sensu-backend
there (using the `install` role which is also part of this same
collection) then it constructs the /etc/sensu/backend.yml configuration
file based on the vars. Eventually it starts backend as a
system service (systemctl or whatever the target OS supports).

Basic integration tests (with default values for parameters) are run
against all supported target platforms. Later, when we enhance the role
with more advanced configuration options, I would suggest, yes, that
we add new scenarios, but it's sufficient that they are run agains one
platform only.

Signed-off-by: Miha Pleško <miha.plesko@xlab.si>