Add support for backend config
This expands the list of supported variables in the role, which
translate into the Sensu backend's configuration file. The role's
variable `backend_config`'s subdocuments will be translated into
the `backend.yml` file, giving the user a wide range of configuration
options. The `backend` role takes care of only a select few of the
configuration items.
We also document the variables that make sense for the backend
according to the Sensu backend documentation.

References:
  * https://docs.sensu.io/sensu-go/5.14/reference/backend/#configuration
  * https://docs.sensu.io/sensu-go/5.14/files/backend.yml