Add metadata to roles
Up until now we had no metadata attached to the individual roles and
galaxy.ansible.com was completely fine with this. But today, things
started failing when we tried to publish a new release of our
collection. ansible-galaxy command reported back this:

  Galaxy import process failed: Role metadata not found (Code: GLW0002)

A quick search through the galaxy sources revealed that this error is
raised if no metadata file can be found in the role directory. And
here we are, adding a fresh set of metadata to our roles.

Does this fix our publish problem? Heck if I know ;) I think it might,
but cannot be sure.