Initialize Sphinx documentation builder
In order to start building the documentation, we bootstrapped the
initial directory structure that Sphinx needs. As promised, we reused
the README.rst file as an entry point.

To install prerequisites, run

    $ pip install -r docs.requirements

Documentation can now be built by running

    $ cd docs && make html

This will place the built documentation into the docs/build/html
directory.