Add support for rendering module documentation
Because manually copying documentation snippets from modules to rst
files does not scale, we introduced ansible-doc-extractor tool to our
documentation pipeline.

Because invoking this tool is not the simplest thing, we also added a
Makefile that contains some sanity checks and commands for rendering
the documentation. With Makefile in place, documentation is just one
`make docs` away.

We also added a module index document that automatically includes all
generated module documents into the main file. This way, module
documentation also shows up in the index.