Add development environment set up documentation
Because setting up a development environment is not exactly
straight-forward, we prepared a short instruction manual for new
developers.

While the document serves mainly as an installation reference, we also
added a few bits of information about the reasons for certain steps
that on the first glance only add complexity to the whole thing.