Split CI pipeline into stages
Up until now we ran all of our jobs in parallel. And while this sound
great in theory, it wastes quite a few resources when one of the
short-running tests fail.

This is why we split the CI jobs into two categories: fast and slow
ones. Now we first run the fast ones and then proceed to run slow
ones. Note that the slow jobs are only run if all of the fast ones
finished successfully.

This saves a ton if resources if one of the fast tests fails.