Make common module arguments composable
Current system of sharing common argument specifications started to
show fundamental deficiencies as we started adding more modules with a
slightly different requirements.

To rectify this situation, we removed the fixed groups of parameter
specifications and replaced them with a single method that can
construct argument specification on demand.

As a welcome side-effect, we now have all module argument names listed
right where we use them: in the module's main function. This does
introduce some duplication into the code base, but the gains are well
worth it.