Flatten integration tests' directory structure
At the moment, our integration tests are all over the place. This
makes it increasingly hard to scale our testing infrastructure.

In this commit, we merged all the tests into a single Molecule
workspace. We moved module scenarios that previously lived under
modules/molecule/<mod> directories into molecule/module_<mod>
directories and scenarios from the roles/<role>/molecule/<s>
directories to the molecule/role_<role>_<s>.

Next, we merged the shared Molecule configuration into a single
base.yml configuration file that lives in the root of the integration
tests. There is no more shared directory under scenarios, making a
mess of our directory structure ;)

To further simplify the test structure, we bumped the Molecule
version. Unfortunately, versions that contain the best stuff are not
released yet, so we need to install molecule directly from the
GitHub's repository.

It may feel a bit weird and scary to go run alpha-quality software.
This means that we will need to track the upstream if anything breaks
and fix the bugs ourselves. And while fixing bugs takes time, sticking
with stable version would mean spending time on herding duplicated
configuration in all of our scenarios.

With that Molecule update behind us, we were able to remove all
Dockerfile templates any symlinks from the repository. This not only
reduced the amount of content in our repositories, it also made sure
that we will not end up with a stale configuration when the updates
inevitably happen.

Another thing that we were able to significantly reduce is the
scenario-specific configuration. Because unstable Molecule allows us
to define common set of platforms right in the shared configuration
file, module tests only need an empty molecule.yml file and that is
it. Unfortunately we cannot remove it completely because Molecule uses
presence of this file as a scenario marker, but still, the saving are
quite significant.