Add optional compare parameter to utils.sync
This commit modifies sync function from the utils module. The
function now accepts an additional argument pointing to the
comparator function to be used when comparing the current and
desired state of objects managed by Sensu modules. The parameter
has a default value pointing to the generic `do_differ` comparator.
With this change, we can support custom comparators tailored for
various Sensu objects, in case the generic one does not suffice.

Signed-off-by: Manca Bizjak <manca.bizjak@xlab.si>