Add Makefiles with helper tasks
While it is necessary to know the development tools in great details,
having a convenience wrappers for running common tasks before
submitting a PR is always welcome.

But the changes in this commit are not only cosmetic. By using make to
execute tasks, we get the parallelism for free. Running a complete
test suite on a local machine previously took almost half an hour. But
with latest make-based tasks, we can execute the whole set of tests in
about ten minutes. And all we need to do is add -j switch to make
invocation.

For example, executing

    make -j8 integration

will run eight scenarios in parallel.