Rewrite README file in markdown
One of the Ansible certification requirements mandates that we use
markdown as a documentation format. This is why we rewrote the README
document in markdown.

But because we would still like to host our more comprehensive
documentation somewhere on the web, we also copied the content into
our root Sphinx document. It is unfortunate that we have to have this
duplication in the codebase, but there is not much we can do about it.