Move validation of role binding module params to role_utils
This commit moves the existing function for validation
of module params from role_binding module to `utils.role_utils`.
We slightly modify the function so that it accepts module params
dict as argument instead of AnsibleModule object. The change is
needed to avoid duplication and allow re-use of the validator
function from the upcoming cluster_role_binding module.

Signed-off-by: Manca Bizjak <manca.bizjak@xlab.si>