Add support for cluster roles in the role_binding module
This commit adds support for specifying cluster role when creating
a role binding. As a result, we're now able to apply cluster roles
in the context of the current namespace. This is different from a
cluster role binding, where a cluster role is applied globally.
The new `cluster_role` parameter is mutually exclusive with `role`.

Signed-off-by: Manca Bizjak <manca.bizjak@xlab.si>