Update event_info module
With this commit, we update event_info module to the new structure.
API supports filtering by entity and by entity and check together,
but does not support filtering by check, if this is needed we can filter
this locally. That's why I created a separate PR for this,
so we can discuss this here. Module is currently not backed with
extensive integration test suite, due to the lack of event module.