Refactor URL path construction
Up until now, we manually constructed paths for Sensu API client. And
while this worked well for us, it is still a bit unsafe and can lead
to some hard-to-debug situations.

In this commit, we introduced the build_url_path function that
performs any escaping while constructing the final path. And while we
were at it, we also made sure this function can handle missing parts
of the path. With this additional functionality, we were able to drop
quite a few if clauses in the code.

We also introduced quoting into the Sensu API client itself since
namespace is another piece of information that we retrieve from the
Ansible user and should be treated with some care.

Note that we did not change the integration tests. This gives us some
confidence that we are on the right track with this change.