Introduce low-level HTTP module
Up until now, the only place where we communicated with the web was in
Sensu client. Now that we are preparing to add a bonsai asset module,
this is no longer true since we need to reach out to bonsai server.

In order to consolidate the common parts, we extracted low-level
functionality from the Sensu client (status code extraction, JSON
parsing) and placed them into a separate module that other HTTP
clients can use.