Add bonsai HTTP client
We created bonsai client because we need to fetch asset versions and
definitions from bonsai.sensu.io. Because this is an external service,
we cannot use Sensu client that we already have.

We tried to keep this client minimalistic, but we did try to validate
data as thoroughly as practically possible, and to create a sensible
error messages for end user.