Add bonsai_asset action plugin
This is an asset module counterpart that knows how to fetch asset
definition directly from the bonsai.sensu.io.

The main reason why we added this plugin is accessibility. Adding a
bonsai asset using good old asset module is really not the most
pleasant experience. The data that we need to provide is not really
hard to obtain, but it is fiddly to properly add to playbook.