Use custom, prebuilt images for integration testing
We used stock docker images for out integration tests, which was not
optimal since we modified each image before running tests on it. But
if we put those inefficiencies aside, things worked great.

But today the builds started failing because official Ubuntu Trusty
image does not work without subscription anymore. This prompted us to
revisit the image situation once again and fix it for good.

What we ended up doing is building our own images that contain
everything tests need to run successfully. They are hosted under the
https://hub.docker.com/u/xlabsi organization on docker hub.

With those images ready for use, we were able to flip the
pre_build_image switch to true in the Molecule's docker driver.

And while we were modifying the image definitions in molecule
definitions, we also added remaining missing versions of Sensu to our
test suite. We now test almost all supported versions of Sensu, with
the sole exception of version 5.15.0 that has a bug in its API.