Make sure agent role works without backends group
In the initial implementation of the automatic agent <-> backend
linking, we assumed that backends group will always be present in the
inventory. Of course, we were completely wrong about that ;)

To address the limitation we just described, we moved the group
selection into the filter. And while this might feel a bit wrong at
the start, it is not all that bad. The filter is already named
backends and highly specialized for a certain task. So nothing tragic
happened here.

Note that we could still solve this problem in the default variable
file using the attr and default filters, but we decided to go with the
customized filter route. This keeps markup minimalistic and all the
logic in the python code where we can unit test it.