Add short functionality descriptions to modules
Our current module documentation told user nothing about what kind of
management functionality it actually offers. Changes in this commit
fix this slightly by explicitly listing the operations that can be
executed (create, update, delete, deactivate, etc.).

And while we were messing around with the documentation, we also
converted the short descriptions into the imperative mood to be in
sync with the official Ansible documentation.