Make info modules return empty list if item is missing
When we initially wrote code info code, we did not give this aspect of
the API much thought. This is why we ended in a situation where

  - asset_info:

and

  - asset_info:
      name: name

return different result when executed against an empty Sensu Go
backend. The first task returns an empty list while the second task
will return a list with a single None in it.

To unify the API, we made the second task also return an empty list.
This also simplifies playbook tests for resource existence since the
playbook authors can check the length of the list and see if there is
anything in it without having to compare list content to None.

We implemented this using a pure filtering function that we inserted
into all info modules. The only exception here is the namespace_info
module that cannot query for a single namespace.

When we added a filtering function, we also made it smart enough to
ignore double-wrapping the lists. Main reason for doing this is to get
rid of the if check in every info module.

Note that we added integration and unit tests. The main purpose of our
unit tests is to make sure our modules have the filter in place while
the integration tests basically check compatibility with the Sensu Go
backend.