Add links to GitHub source files
Because our documentation is not perfect by any means, we added source
links to documentation pages. This should make it easier for people
who find an error in our docs to prepare an error report and maybe
even submit a fix.

We took special care to link auto-generated documentation to the
relevant sources. Currently, this means that module reference
documentation points directly to the module source code where the
documentation lives.