Prepare documentation for the 1.0 release
We added or modified all of the vital parts of the documentation in
preparation for the 1.0 release. Note that since documentation is not
hosted on Ansible Galaxy or Automation Hub, we can easily improve it
even after the release as long as we do not remove existing sections
and break the documents that link to our docs.

Co-Authored-By: mancabizjak <manca.bizjak@xlab.si>