Remove flaky tests that relied on external data
When we were determining the latest released version, we used an ugly
hack that relied on docs.sensu.io redirect to hold the right version
information.

Because Sensu changed the documentation site, our tests started to fail.
And we have no one but ourselves to blame for that one.

We removed the ugly tests from the repository and while this reduces
the comprehensiveness of the tests, we cannot do better right now
since there is no canonical place to look for the latest version.