Simplify component version selection
Because we tightly coupled Sensu Go components with their versions, we
ended up in a situation where it was almost impossible to install
anything but the latest version of components.

Why? For example, our backend role sets the components variable to
sensu-go-backend. If playbook author would like to install some older
version of backend, she would need to provide new value for components
variable, possibly breaking our backend role.

To get out of this mess, we added a new variable to the install role.
We called it version for obvious reasons. And this way we are one step
closer to untangling our somewhat messy roles ;)