Add external datastore modules
Datastore module and its info counterpart know how to deal with
external datastores.

Payload that the Sensu storage API accepts is a bit different from the
payload that core/v2 API endpoints expect, but we kept the user-facing
interface similar to all other modules. This way, playbook authors can
use their knowledge about existing modules when dealing with data
storages.

Another peculiarity with datastore module is its reluctance to break
the backend. This means that datastore module will abort adding
external datastore if one is already found on the backend. Note that
it is still possible to remove and update datastores even if there is
more than one on the backend.