Add flake8 pass to sanity tests
When we upload our collection to Ansible Galaxy, the galaxy backend
performs a static analysis of the submitted source code and uses it to
calculate the quality score of the collection.

Unfortunately, the checks that galaxy backend runs are not documented
anywhere. But what we do get is the log of the import process that
contains some clues about the tests.

This commit adds a flake8 check to the group of sanity tests. The
configuration for the flake8 was extracted directly from the galaxy
source code and tweaked a bit for the use in our repository.