Fix bonsai_asset parameter validation on python 2
When Ansible is running on python 2, it converts all strings into
unicode type when loading YAML files. And because our validation
function did not account for this, all hell break loose on python 2.

This commit introduces a simple fix for this: instead of checking for
a fixed type, we use Ansible-provided python-version-specific wrappers
for a string type.

Note that we could also use unicode type directly, but since Ansible
developers already maintain a cross-version implementation, we might
as well use it.