Strip empty lines from Sensu configuration files
While it is true that the generated configuration files should not be
modified manually, they might still get inspected from time to time.
And in those cases, having a nice, readable configuration file can
make a significant difference in usability.

But because we do not want to bundle all our template conditional
logic into one continuous block that looks like a core dump, we
implemented the cleanup by using Jinja's trim modifiers to code
blocks.

This way we can almost have a cake and eat it too: the template is
readable from developer's perspective and the generated configuration
is concise and clean.