Add package_name filter
The newly added filter's task is to group together the knowledge about
the distribution-specific package name conventions. This will allow us
to keep the roles free of multitude of conditionals that we would need
when constructing file names.

At the moment, filter only supports apt and yum tools, but adding
additional package managers should be relatively easy in the future
(for example, when we will add support for dnf that is used in RHEL 8).