Remove hardcoded apt key id
Since keys are channel-specific, it does not make any sense to have a
single key id hardcoded into our role. We do not hardcode yum keys, so
there is no real reason to hardcode them for apt.

There is one thing that we loose with the id removal: the ability to
determine the changes in check mode. But this should have almost no
effect in the real-world scenarios where repository setup is not a
primary goal of using the install role.