Add build support to the install role
Up until now, our install role knew nothing about the build
specifiers. In practice, this meant that it was not possible to
update the specific version of the package to the latest build.

In order to rectify this, we added another variable to the install
role that allows playbook authors to pin down the package version
completely. This should make system admins that strive for
reproducible deploys very happy.

We also changed the semantics of the install role a bit. First,
in the simplest variant, we can have a role invocation like this:

  - include_role:
      name: sensu.sensu_go.install
    vars:
      components: [sensu-go-agent]

When this task is executed, it will install or update sensu-go-agent
package to the latest currently available version and build. If we add
a version specifier, we get this invocation:

  - include_role:
      name: sensu.sensu_go.install
    vars:
      components: [sensu-go-agent]
      version: 5.16.0

This will install latest available build of version 5.16.0. And to
fully pin down the version, we can write

  - include_role:
      name: sensu.sensu_go.install
    vars:
      components: [sensu-go-agent]
      version: 5.15.0
      build: 8290

Note that downgrading packages is not supported.