Make sure we run all integration tests
Because our CI runs only tests that are listed in the file with
scenario run times, we need to make sure that when new test is
created, it is also timed and timings committed.

The check for this is really simple: we list all of the scenarios in
the molecule directory and compare the output with the contents of the
timings file. If there is anything different, we abort the run and
fail the test.