Fix apt package installation on fresh instances
Up until now, installation role assumed that apt's package cache is at
least partially up to date. It turned out that this assumption fails
flat on its face in the world of cloud images.

Changes in this commit force a cache refresh before any of the
preparation processes are started. It may seem a bit expensive to do
this at first, but the performance impact is actually negligible since
the update happens anyway when we add a new repo.