Replace installation role tests with playbook checks
Changes in this commit unify our test suite and standardize all of
them into playbook checks. The main reason for having checks in the
playbook is the ability to test the idempotency of the operations and
the convenience of having tests right next to the action that we are
testing.