Add API key authentication support
Sensu Go supports API key authentication since version 5.15.0 and it
is about time we also add it as a valid method of authentication.

In order not to break the backwards compatibility, we made the API key
optional. If the API key is not present in the parameters, we proceed
with the username and password login. If the API key is set, we
validate it and in the case of failure report the error back to the
user.

Note that if the API key is present nut invalid, we do not try to
authenticate using username and password. We wrote the code this way
on purpose, since silently ignoring authentication failure with
something that user believes is valid is not an option.