Add check that validates supported role platforms
Galaxy performs some tests when importing roles. One of them is
checking for a validity of the supported platforms. And because no
existing tool does this for us, we created a simple python script that
queries the galaxy web API and ensures that we have a valid set of
platforms listed in role metadata.