Pin CircleCI docker images
One of the image updates broke our CI. And because the error message
we get on failure is almost completely useless, it will take some time
to get to the bottom of this issue.

And while we cannot completely bypass this problem (we must update our
test images from time to time), we can reduce its impact on the
development team by pinning the images down to the SHA256 hash. This
way, we can perform the image updates ,under our terms and conditions,
but more importantly, we remove one moving part from the CI pipeline.