Add support for CentOS/RHEL 8
The latest version of the OS from Red Hat uses dnf to manage system
packages. Fortunately, Ansible already contains appropriate modules so
adding support was relatively painless.

Do note that we copied the contents of the prepare.yml file from the
yum repo verbatim. We did this on purpose because we do not want dnf
and yum functionality to be tied together unnecessarily.