Remove tags from our roles
Using tags in roles is almost useless since the only way to control
what gets executed is by using the --[skip-]tags Ansible parameter.
We cannot store this information in a configuration file or playbook
itself, which makes it rather useless.

Since all of our roles are idempotent, there is little need to skip
certain tasks when executing a play. Adding a few seconds to the play
runtime should not matter much in vast majority of cases.

There is also one very ugly side to our tags usage: it allows users to
break their runs. For example, package installation will fail if the
user instructed Ansible to skip the repository setup step.

We never documented the tags, so this change does not break the public
API of the roles.