Remove Ubuntu 19.04 from integration tests
Ubuntu 19.04 has reached its end of life. Because most of the apt
repositories stopped working, it would be pretty hard to get
installation working again. And all of that effort would be wasted
since it is impossible to even buy support for Ubuntu 19.04.

DO note that we did not remove the support for this Ubuntu version
from the collection, we just stopped testing it.