Split backend tasks into separate files
We decided that we split the backend role into three parts:
installation, configuration, and start. This way, we can run each set
of tasks independently.

This split did introduce a minor complication though. Our backend
restart handler now needs to be executed conditionally. But we made
sure that the behavior is backwards-compatible with the previous
versions.