Make sure check mode works on backend role
Because we are using the command module to test the `sensuctl-backend
init` command presence, we need to handle check mode a bit
differently. The main reason is that the command module does not
support check mode when we do not supply the creates or removes
parameter.

Luckily for us, displaying help does not change the system at all so
we can run this command safely in check mode.