Extend do_differ function
The function in question now takes an optional list of keys that
should be ignored when comparing the payload. This should make it
easier to use this function for partial comparisons in modules that
contain some custom detection logic for certain keys.

The second change is related to the metadata handling. Because Sensu
Go started adding auto-generated created_by field to the metadata,
comparison function learned how to ignore that parameter when
comparing data.