Lift the check for non-existent remote object
Comparison method was up until now equipped with enough logic to
handle the missing remote objects. But this is quite unfortunate since
there are quite a few custom comparison implementations in our code
base and they all need their own copy of this logic.

In order to reduce the duplication, we moved the logic from the
comparison method into to the synchronization function. This allowed
us to drop quite a few if checks throughout the codebase.