Ensure custom comparators properly handle metadata
Up until now, custom comparators usually copied the default comparator
logic and then sprinkled custom processing on top. But since the
default comparator gained an ignore list, we can delegate most of the
work directly to the base comparator and only perform custom work.

Main reason for doing this is addition of the metadata handling. If we
kept the current state, we would need to copy that logic into all
custom comparators.