Update the set of backend versions we test against
Running tests against all minor versions is a bit of an overkill and
wastes quite a lot of time and energy for only marginal benefit. This
is why we rimmed down the list of versions that we test.

Under the new regime, we test only the latest patch release in each
supported minor release.