Fix module documentation
Preliminary run of `ansible-test sanity` uncovered some errors that we
had in our documentation. We mostly forgot to specify the types of
elements in a list, which was rather trivial (and somewhat boring)
error to fix.

We also took a chance and made sure our argument specification matches
the documentation.