Use FQCNs in examples
In order to reduce the cognitive load on users that have little or no
experience with Ansible and Ansible Collections, we replaced short
module names with a fully-qualified ones. This way, users can
copy-paste examples directly without having to prefix the names with
sensu.sensu_go or use `collections: [ sensu.sensu_go ]` directive.