Make user module idempotent
Existing user module implementation had quite a few shortcomings that
were limiting its the usefulness. The main problem was related to the
password handling: because the backend does not return passwords back
when we query for user data (which is how things should be), we had no
way of testing if the credentials matched, so we always returned
`changed: true` when encountering a password.

With the addition of authentication check call, we were able to remove
this limitation and make module idempotent. We do need to make an
additional API call for this, but this should not matter much since
mass-editing users is not a realistic scenario.

And while we were fixing the (non)idempotency of the module, we also
removed the condition that the groups parameter also requires password
parameter to be set. Now all of the parameter are orthogonal to each
other and can be set independently.