Add docker definitions to repo
Up until now, we kept the docker definitions in a separate repository,
but that made it quite hard to keep the things in sync.

Changes in this commit move the docker image definitions into the
collection. In order to prevent any name clashes, we also prefix all
image names with the `sensu-go-tests-`, explicitly stating what those
images are used for.