Get datastore-related modules back in sync with the backend
Datastore API changed a bit somewhere between 5.15.0 and 5.18.0. In
order to keep up with the changes, we updated the module code in order
to handle the change.

Note that we did try to make the change backwards-compatible, but
since the versions of the Sensu Go that contained an old API are not
supported anymore by the upstream, we do not test for this
compatibility.