Make sure the check module is idempotent
Payload for creating a check is large and contains quite a few lists
that serve as a set of things. All this makes comparing the payloads a
bit tougher that for the rest of the resources.

Another source of problems is the backend's tendency to treat null and
empty list as equivalent, adding yet another place where equality
checks can go wrong.

Changes in this commit try to fix all of the idempotency issues we
could think of, but there is a slight possibility that we still missed
something. But things should be in a much better shape now.