Adapt user module for Sensu Go 5.21.0 and newer
Sensu Go 5.21.0 broke the backwards-compatibility of the user API.
Changes in this commit gloss over those changes and make sure playbook
authors can still use the same API as before.

Note that because the backend now accepts an encrypted password in the
payload, the collection gained another dependency that needs to be
installed on the node that module executes on: bcrypt.