Add options for customizing the certificate validation
Up until now, the HTTP client that we use to interact with the Sensu
Go backend always verified the backend certificate using the same CAs
as the python itself.

In this commit, we added two new parameters to the authentication part
of the module. The verify parameter allows users to disable the
verification altogether and should only be used in development
environments. In production, users should either use certificates that
are trusted by default (via systems root CAs) or use a ca_path
parameter that allows users to specify what CA bundle should be used
to validate the backend certificate.

Both new parameters can also be set via the environment variables.
We added this fallback option to keep the auth module parameter
optional as it is now.