Add sanity test ignore file for Ansible 2.9
Our collection does not support Python 2.6 because that version is
long obsolete and almost impossible to test against. Unfortunately for
us, the Automation Hub import process still runs the compile tests
using that version.

To avoid getting errors during the import process, we added an ignore
file that instructs ansible-test to skip the compilation test on
certain files that use constructs introduces in Python 2.7.